+++ packages/contracts/contracts/honeylemon/MinterBridge.sol
@@ -1,21 +1,21 @@
 pragma solidity 0.5.2;
 
-import 'openzeppelin-solidity/contracts/ownership/Ownable.sol';
-import 'openzeppelin-solidity/contracts/token/ERC20/SafeERC20.sol';
-import 'openzeppelin-solidity/contracts/token/ERC20/ERC20.sol';
-import 'openzeppelin-solidity/contracts/utils/ReentrancyGuard.sol';
+import "openzeppelin-solidity/contracts/ownership/Ownable.sol";
+import "openzeppelin-solidity/contracts/token/ERC20/SafeERC20.sol";
+import "openzeppelin-solidity/contracts/token/ERC20/ERC20.sol";
+import "openzeppelin-solidity/contracts/utils/ReentrancyGuard.sol";
 
-import '../libraries/MathLib.sol';
-import '../marketprotocol/MarketCollateralPool.sol';
-import '../marketprotocol/mpx/MarketContractMPX.sol';
-import '../marketprotocol/tokens/PositionToken.sol';
+import "../libraries/MathLib.sol";
+import "../marketprotocol/MarketCollateralPool.sol";
+import "../marketprotocol/mpx/MarketContractMPX.sol";
+import "../marketprotocol/tokens/PositionToken.sol";
 
-import './MarketContractProxy.sol';
+import "./MarketContractProxy.sol";
 
 
 contract MinterBridge is ReentrancyGuard, Ownable {
-    using MathLib for uint;
-    using MathLib for int;
+    using MathLib for uint256;
+    using MathLib for int256;
     using SafeERC20 for ERC20;
 
     MarketContractProxy public marketContractProxy;
@@ -31,7 +31,7 @@ contract MinterBridge is ReentrancyGuard, Ownable {
      * @notice check that called is 0x minter bridge proxy address
      */
     modifier only0xBridgeProxy() {
-        require(msg.sender == ERC20_BRIDGE_PROXY_ADDRESS, 'invalid caller');
+        require(msg.sender == ERC20_BRIDGE_PROXY_ADDRESS, "invalid caller");
         _;
     }
     /**
@@ -40,7 +40,7 @@ contract MinterBridge is ReentrancyGuard, Ownable {
     modifier onlyIfSetMarketContractProxy() {
         require(
             MARKET_CONTRACT_PROXY_ADDRESS != address(0),
-            'MarketContractProxy not set'
+            "MarketContractProxy not set"
         );
         _;
     }
@@ -95,12 +95,14 @@ contract MinterBridge is ReentrancyGuard, Ownable {
         // The proxy acts as the taker token to make 0x think that the appropriate amount
         // was transferred and accept the trade as passing. Under the hood the  marketContractProxy
         // has minted long and short tokens and sent them to the the investor and miner.
-        require(tokenAddress == MARKET_CONTRACT_PROXY_ADDRESS, 'bad proxy address');
+        require(tokenAddress == MARKET_CONTRACT_PROXY_ADDRESS, "bad proxy address");
 
-        // (imBTC) sent from the miner
+        // (wBTC) sent from the miner
         ERC20 collateralToken = ERC20(marketContractProxy.COLLATERAL_TOKEN_ADDRESS());
 
-        uint neededCollateral = marketContractProxy.calculateRequiredCollateral(amount);
+        uint256 neededCollateral = marketContractProxy.calculateRequiredCollateral(
+            amount
+        );
 
         collateralToken.safeTransferFrom(from, address(this), neededCollateral);
         collateralToken.approve(MARKET_CONTRACT_PROXY_ADDRESS, neededCollateral);
