+++ packages/contracts/contracts/marketprotocol/tokens/MarketToken.sol
@@ -16,7 +16,7 @@
 
 pragma solidity 0.5.2;
 
-import './UpgradableToken.sol';
+import "./UpgradableToken.sol";
 
 
 /// @title Market Token
@@ -24,11 +24,11 @@ import './UpgradableToken.sol';
 /// as well as have a minimum balance of tokens to create new Market Contracts.
 /// @author Phil Elsasser <phil@marketprotocol.io>
 contract MarketToken is UpgradeableToken {
-    string public constant name = 'MARKET Protocol Token';
-    string public constant symbol = 'MKT';
+    string public constant name = "MARKET Protocol Token";
+    string public constant symbol = "MKT";
     uint8 public constant decimals = 18;
 
-    uint public constant INITIAL_SUPPLY = 600000000 * 10**uint(decimals); // 600 million tokens with 18 decimals (6e+26)
+    uint256 public constant INITIAL_SUPPLY = 600000000 * 10**uint256(decimals); // 600 million tokens with 18 decimals (6e+26)
 
     constructor() public {
         _mint(msg.sender, INITIAL_SUPPLY);
